Training Script Complete
This script is intended to run on google colab, to train the summarizer model with the defined parameters (max length of summary: 150, set to not save intermediate checkpoints to save storage, save only the best model, train for 25 epochs) etc. and save the model to google drive as a zip file for download and local use.